Move test util into internal dir
